[DM-34433] Stop controlling CORS headers
This is the code that used to set the CORS headers for incoming
requests.  This doesn't seem to work anymore on kubernetes and
ingress-nginx that we recently updated to, which tries to control
the headers itself, but does allow strips the headers from the
service.  So to make this less confusing, stop trying to use this
here as the headers will be stripped anyway.